requirements.txt
pywinauto
pyautogui
pynput
pillow